+++ src/game/client/ui/ms/vgui_menu_interact.h
@@ -46,8 +46,8 @@ public:
 
 		m_ButtonY = BTN_START_Y;
 
-		for (int i = 0; i < INTERACT_MAX_BUTTONS; i++) //Create all the buttons now.  Fill them in later
-			AddButton("", 0, i);
+		for (unsigned int i = 0; i < INTERACT_MAX_BUTTONS; i++) //Create all the buttons now.  Fill them in later
+			AddButton("", 0, (int)i);
 
 		// MiB NOV2014_25, center the title and separator NpcInteractMenus.rft
 		m_AllowKeys = true; // MiB 25NOV_2014 - Disable key-input
@@ -63,7 +63,7 @@ public:
 		m_EntIdx = -1;
 		m_Options.clearitems();
 
-		for (int i = 0; i < m_Buttons.size(); i++) //Reset all buttons except cance
+		for (unsigned int i = 0; i < m_Buttons.size(); i++) //Reset all buttons except cance
 		{
 			m_Buttons[i]->setVisible(false);
 			m_Buttons[i]->setEnabled(true);
